tidy up aws_scripts
